fix: address thread review feedback
